update readme & code refactoring.
